requirements.txt
elasticsearch==8.11.0
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
nltk==3.8.1
scikit-learn==1.3.2
Pillow==10.1.0 